feat: finish role-based access for vacancy CRUD, add method   findAllVacanciesDetailed() accessible only for superAdmin, admin and recruiter

+++ src/vacancy/vacancy.service.ts
@@ -8,6 +8,7 @@ import { vacancyToVacancyDto } from 'src/vacancy/map/vacancy.map';
 import { CreateVacancyDto } from 'src/vacancy/dto/createVacancy.dto';
 import { UpdateVacancyDto } from 'src/vacancy/dto/updateVacancy.dto';
 import { UserDto } from 'src/user/dto/user.dto';
+import { validateTenantAccess } from 'src/utils/validate';
 
 @Injectable()
 export class VacancyService {
@@ -17,10 +18,15 @@ export class VacancyService {
   ) {}
 
   async findAll(): Promise<VacancyDto[]> {
+    const vacancies = await this.vacancyRepository.find();
+    return vacancies.map(vacancyToVacancyDto);
+  }
+
+  async findAllDetailed(): Promise<VacancyDto[]> {
     return await this.vacancyRepository.find();
   }
 
-  async findByVacancyId(vacancyId: string): Promise<VacancyDto> {
+  async findDtoByVacancyId(vacancyId: string): Promise<VacancyDto> {
     const vacancy = await this.vacancyRepository.findOne({
       where: { id: vacancyId },
     });
@@ -32,7 +38,7 @@ export class VacancyService {
     return vacancyToVacancyDto(vacancy);
   }
 
-  async findByTenantId(tenantId: string): Promise<VacancyDto[]> {
+  async findAllByTenantId(tenantId: string): Promise<VacancyDto[]> {
     const tenant = await this.vacancyRepository.find({
       where: { tenantId },
     });
@@ -47,13 +53,13 @@ export class VacancyService {
     const vacancies = await this.vacancyRepository.find({
       where: { tenantId },
     });
-    return vacancies.map(vacancyToVacancyDto);
+    return vacancies;
   }
 
   async create(
     createVacancyDto: CreateVacancyDto,
     creator: UserDto,
-  ): Promise<VacancyDto> {
+  ): Promise<Vacancy> {
     const vacancy = this.vacancyRepository.create(createVacancyDto);
 
     if (creator.tenantId) vacancy.tenantId = creator.tenantId;
@@ -66,17 +72,35 @@ export class VacancyService {
   async update(
     vacancyId: string,
     updateVacancyDto: UpdateVacancyDto,
+    updater: UserDto,
   ): Promise<VacancyDto> {
-    const vacancy = await this.findByVacancyId(vacancyId);
+    const vacancy = await this.findVacancyByVacancyId(vacancyId);
+    validateTenantAccess(updater, vacancy.tenantId);
 
     Object.assign(vacancy, updateVacancyDto);
 
+    vacancyToVacancyDto(vacancy);
     return await this.vacancyRepository.save(vacancy);
   }
 
-  async remove(vacancyId: string): Promise<VacancyDto> {
-    const vacancy = await this.findByVacancyId(vacancyId);
+  async remove(vacancyId: string, deleter: UserDto): Promise<VacancyDto> {
+    const vacancy = await this.findVacancyByVacancyId(vacancyId);
+    validateTenantAccess(deleter, vacancy.tenantId);
+
     await this.vacancyRepository.delete(vacancyId);
+
+    vacancyToVacancyDto(vacancy);
+    return vacancy;
+  }
+
+  private async findVacancyByVacancyId(vacancyId: string): Promise<Vacancy> {
+    const vacancy = await this.vacancyRepository.findOne({
+      where: { id: vacancyId },
+    });
+
+    if (!vacancy) {
+      throw new HttpException('Vacancy is not found.', HttpStatus.NOT_FOUND);
+    }
 
     return vacancy;
   }
